Add notion of 'skipped' tests to rtest.py
The new --skipped option replaces the --all-tests option. Test
discovery now includes all tests by default, and --skipped filters it
so only skipped tests (whether by virtue of being marked as skipped or
being incomplete) are returned. This makes the --list option a bit
more useful; rtest.py --list --skipped --verbose will show the user
all tests that are skipped and for which reason. When regression tests
are run, skipped tests (for whatever reason) are printed with the
result "SKIP" and are counted as with PASS, FAIL, and ERROR.