Populate TestByGeneration page as much as possible
This is a small change but it will enable us to view everything the test by generation page is supposed to display excluding any missing information due to issues with generating the sentences.

In particular, if generation was unsuccessful the sentences, parse trees, and mrses will be missing but you can still see what patterns were tried and what predications were selected.

This way you know the system has tried to generate sentences but no sentences were generated.