Enable Generate from Test Sentences
This change will allow users to generate sentences using the Test Sentences in the choices file.

Workflow
1. In the Test by Generation Options page, select 'Generate From Test Sentences'
This will add `from-test-sentences=on` to the choices file under the `gen-options` section
2. Save the choices file then navigate to 'Test by Generation' page
3. Test By Generation parses the test sentences and displays each sentence
Note: The generated sentence is *exactly* the same as the test sentence because no changes were made to the MRS; however, you can now see the MRS and parse tree of the test sentence. If it doesn't parse or there are no test sentences at all, you will not see the test sentences at all. In the future it may be helpful to at least return a message that the parse failed.
4. If you select `More sentences with this verb and pattern` you can see more sentences with the same verb but different determiners/nouns. Note that if the sentence was specific about person, number, gender etc. then those constraints will also apply to the new sentences.